src/requirements.txt
facenet_pytorch==2.5.2
mediapipe==0.9.0.1
numpy==1.23.5
opencv_python==4.6.0.66
Pillow==9.5.0
torch==2.0.0
